move modeldb-sqlite-shared into a subpath export of modeldb
