+++ packages/modeldb-durable-objects/package.json
@@ -14,7 +14,6 @@
 	"sideEffects": false,
 	"dependencies": {
 		"@canvas-js/modeldb": "0.11.0",
-		"@canvas-js/modeldb-sqlite-shared": "0.11.0",
 		"@canvas-js/utils": "0.11.0",
 		"@cloudflare/workers-types": "^4.20240925.0",
 		"@ipld/dag-json": "^10.2.2",
