+++ packages/modeldb-durable-objects/tsconfig.json
@@ -6,5 +6,5 @@
 		"outDir": "lib",
 		"types": ["@cloudflare/workers-types"]
 	},
-	"references": [{ "path": "../modeldb" }, { "path": "../modeldb-sqlite-shared" }, { "path": "../utils" }]
+	"references": [{ "path": "../modeldb" }, { "path": "../utils" }]
 }
