--- packages/modeldb-sqlite-shared/tsconfig.json
@@ -1,9 +1,0 @@
-{
-	"extends": "../../tsconfig.base.json",
-	"include": ["src"],
-	"compilerOptions": {
-		"rootDir": "src",
-		"outDir": "lib"
-	},
-	"references": [{ "path": "../interfaces" }, { "path": "../modeldb" }, { "path": "../utils" }]
-}
