+++ packages/modeldb-sqlite-wasm/package.json
@@ -18,7 +18,6 @@
 	"dependencies": {
 		"@canvas-js/interfaces": "0.11.0",
 		"@canvas-js/modeldb": "0.11.0",
-		"@canvas-js/modeldb-sqlite-shared": "0.11.0",
 		"@canvas-js/utils": "0.11.0",
 		"@libp2p/logger": "^5.1.0",
 		"@sqlite.org/sqlite-wasm": "^3.46.0-build2",
