+++ packages/modeldb-sqlite-wasm/tsconfig.json
@@ -5,10 +5,5 @@
 		"rootDir": "src",
 		"outDir": "lib"
 	},
-	"references": [
-		{ "path": "../interfaces" },
-		{ "path": "../modeldb" },
-		{ "path": "../modeldb-sqlite-shared" },
-		{ "path": "../utils" }
-	]
+	"references": [{ "path": "../interfaces" }, { "path": "../modeldb" }, { "path": "../utils" }]
 }
