+++ packages/modeldb-sqlite/package.json
@@ -17,7 +17,6 @@
 	},
 	"dependencies": {
 		"@canvas-js/modeldb": "0.11.0",
-		"@canvas-js/modeldb-sqlite-shared": "0.11.0",
 		"@canvas-js/utils": "0.11.0",
 		"@types/better-sqlite3": "^7.6.10",
 		"better-sqlite3": "^11.0.0"
