+++ packages/modeldb/src/utils/encoding.ts
