+++ packages/modeldb/tsconfig.json
@@ -1,6 +1,6 @@
 {
 	"extends": "../../tsconfig.base.json",
-	"include": ["src"],
+	"include": ["src", "src/sqlite-utils"],
 	"compilerOptions": {
 		"rootDir": "src",
 		"outDir": "lib"
