+++ tsconfig.json
@@ -23,7 +23,6 @@
 		{ "path": "./packages/modeldb-idb" },
 		{ "path": "./packages/modeldb-pg" },
 		{ "path": "./packages/modeldb-sqlite" },
-		{ "path": "./packages/modeldb-sqlite-shared" },
 		{ "path": "./packages/modeldb-sqlite-wasm" },
 		{ "path": "./packages/chain-atp" },
 		{ "path": "./packages/chain-atp/test" },
